edited: database migrations

+++ backend/db/migrations/universities/000004_create_specialities_table.up.sql
@@ -4,7 +4,7 @@ CREATE TABLE if not exists universities.specialities(
     id SERIAL PRIMARY KEY,
     university_id INT REFERENCES universities.universities(id),
     name TEXT NOT NULL,
-    budget_points FLOAT,
-    contract_points FLOAT,
+    budget_points INT,
+    contract_points INT,
     cost INT NOT NULL
 )
